model/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
keras==2.9.0
numpy==1.22.4
tensorflow==2.9.1
